Implemented ConsoleInterface.
Implemented ConsoleGlyph.
 -ConsoleWriteable interface
 -AbstractGameAsset
 -package restructure: character, items, world moved from main to asset
