+++ src/console/Canvas.java
@@ -1,4 +1,4 @@
-package main;
+package console;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
