+++ src/console/ConsoleTile.java
@@ -1,4 +1,4 @@
-package main;
+package console;
 
 import java.awt.image.BufferedImage;
