+++ src/console/DisplayPanel.java
@@ -1,4 +1,4 @@
-package main;
+package console;
 
 import javax.swing.*;
 import java.awt.*;
